Move case announcements over into nsw-covid.js file
